pyproject.toml
[project]
name = "smoke-watch"
version = "0.1.0"
description = "Air quality monitoring pipeline for MLOps portfolio"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "pyarrow>=23.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.14.13",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "joblib>=1.3.2",
    "pyyaml>=6.0.3",
]

[project.scripts]
fetch-data = "data.fetch_data:main"
check-quality = "data.check_quality:check_raw_data"
    
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pyyaml>=6.0.3",
    "nbconvert>=7.16.6",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "requests-mock>=1.12.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "I"]
# Redirect ruff cache to a hidden folder to keep root clean
cache-dir = ".cache/ruff"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
# Redirect pytest cache to a hidden folder
cache_dir = ".cache/pytest"

[tool.coverage.run]
# Redirect coverage data
data_file = ".cache/coverage/.coverage"
